Update README.md
This commit introduces a minor but important adjustment in the README documentation to enhance clarity regarding service updates. Specifically, we have moved the note about the web3.storage upload API sunsetting to a new, separate line below the list of account setup links. This change ensures there is no confusion that the sunsetting of the upload API is related specifically to Filebase. The intention is to clearly differentiate between the services listed and the general update about web3.storage's API transition.